consolidated utility files, renamed script that determines required transformations
